# Conflicts:
#	build.gradle
#	src/main/java/com/example/instaclone/domain/comment/controller/CommentController.java
#	src/main/java/com/example/instaclone/domain/comment/dto/CommentRequestDto.java
#	src/main/java/com/example/instaclone/domain/comment/entity/Comment.java
#	src/main/java/com/example/instaclone/domain/comment/repository/CommentRepository.java
#	src/main/java/com/example/instaclone/domain/comment/service/CommentService.java
#	src/main/java/com/example/instaclone/domain/like/LikeService.java
#	src/main/java/com/example/instaclone/domain/like/Likes.java
#	src/main/java/com/example/instaclone/domain/post/controller/PostController.java
#	src/main/java/com/example/instaclone/domain/post/dto/PostRequestDto.java
#	src/main/java/com/example/instaclone/domain/post/dto/PostResponseDto.java
#	src/main/java/com/example/instaclone/domain/post/dto/PostResponseDtoImpl.java
#	src/main/java/com/example/instaclone/domain/post/entity/Post.java
#	src/main/java/com/example/instaclone/domain/post/entity/Timestamped.java
#	src/main/java/com/example/instaclone/domain/post/repository/PostRepository.java
#	src/main/java/com/example/instaclone/domain/post/service/PostService.java
#	src/main/java/com/example/instaclone/domain/user/controller/UserController.java
#	src/main/java/com/example/instaclone/domain/user/dto/CheckEmailRequestDto.java
#	src/main/java/com/example/instaclone/domain/user/dto/MyPageResponseDto.java
#	src/main/java/com/example/instaclone/domain/user/dto/SignupRequestDto.java
#	src/main/java/com/example/instaclone/domain/user/entity/User.java
#	src/main/java/com/example/instaclone/domain/user/repository/UserRepository.java
#	src/main/java/com/example/instaclone/domain/user/service/KakaoService.java
#	src/main/java/com/example/instaclone/domain/user/service/UserService.java
#	src/main/java/com/example/instaclone/global/aws/AwsS3UploadService.java
#	src/main/java/com/example/instaclone/global/aws/S3Component.java
#	src/main/java/com/example/instaclone/global/aws/UploadService.java
#	src/main/java/com/example/instaclone/global/config/WebSecurityConfig.java
#	src/main/java/com/example/instaclone/global/exception/GlobalExceptionHandler.java
#	src/main/java/com/example/instaclone/global/jwt/JwtAuthFilter.java
#	src/main/java/com/example/instaclone/global/security/UserDetailsImpl.java

+++ src/main/java/com/example/instaclone/domain/comment/service/CommentService.java
@@ -10,7 +10,6 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-
 @RequiredArgsConstructor
 @Service
 public class CommentService {
@@ -21,7 +20,7 @@ public class CommentService {
 
     @Transactional
     public void createComment(Long postId, CommentRequestDto commentRequestDto, User user) {
-//        게시글 존재 여부 확인. 없으면 예외처리
+//      게시글 존재 여부 확인. 없으면 예외처리
         Post post = postRepository.findById(postId).orElseThrow(
                 () -> new IllegalArgumentException("게시글을 찾을 수 없습니다.")
         );
@@ -31,17 +30,21 @@ public class CommentService {
         post.getComments().add(comment);
     }
 
-    //    2. 댓글 삭제 메서드
+    //  댓글 삭제 메서드
     @Transactional
     public void deleteComment(Long postId, Long commentId , User user) {
         //        게시글 존재 여부 확인. 없으면 예외처리
         Post post = postRepository.findById(postId).orElseThrow(
                 () -> new IllegalArgumentException("게시글을 찾을 수 없습니다.")
         );
-//        댓글 존재 여부 확인. 없으면 예외처리
+//      댓글 존재 여부 확인. 없으면 예외처리
         Comment comment = commentRepository.findById(commentId).orElseThrow(
                 () -> new IllegalArgumentException("댓글을 찾을 수 없습니다.")
         );
+        // 내가 쓴 코멘트인가 확인
+        if(!user.getId().equals(comment.getUser().getId()))
+            throw new IllegalArgumentException("사용자가 일치하지 않습니다.");
+
         post.commentCountMinus();
         commentRepository.deleteById(commentId);
     }
